mark ec2 instance as unhealth if user-data fails
Sometimes the cloud-init / user-data boot script can fail to
successfully bootstrap an ECS cluster node (failure to start agent,
network not ready yet, phase of moon wrong for docker to work).
Currently this leaves an orphened EC2 instance that cannot recover and
requires human intervention to terminate.

This change traps errors in the script execution and triggers a call to
the set-instance-health api to mark the instance as "Unhealth", this
will allow the asg's replacement logic to retire the instance and try
again itself.