egress-proxy: switch to using the fargate version
We currently have two egress-proxy services deployed (one on ECS/EC2 and
one on ECS/Fargate).

The EC2 version uses a Classic ELB to forward traffic to tasks.

The fargate version has service discovery enabled so that traffic can be
sent directly to the running tasks.

This changes the saml-engine tasks to use the addresses of the
discovered fargate tasks and removes the use of the egress proxy from
the remaining ec2 clusters (which both have open egress to 80/443 anyway).