HUB-738: Refactor metadata to separate module
This creates a new module, metadata, which is mostly a duplicate of the
`hub_metadata` config within the hub module.

The point of this is to allow the metadata terraform to be applied
separately from the hub terraform - the ultimate goal being a separate
build/deploy pipeline for the metadata rather than being part of the hub
pipeline.

This also adds some config to the hub module to allow the metadata
config to be "turned off" on an environment basis. This is to allow
migration of the metadata to the new pipeline by environment. The
default specified is maintain the current behaviour, making these
changes a no-op.

A few more outputs have also been specified in the hub module - this so
they can be picked up by the new metadata module via the `remote_state`
data.

There will be a separate PR on the verify-infrastructure-config repo
that will invoke the new metadata module as well as configure the
current hub module to stop managing the metadata infrastructure per
environment.

We want to move the management of the metadata infrastructure to the new
module without ever destroying what's currently running. The envisaged
method to do this is (per environment):

    1. Use `terraform state rm` to remove the metadata resources from
       the hub module's state file. They'll still be running, terraform
       will just no longer be aware of them.
    2. Update verify-infrastructure-config repo to set `manage_metadata`
       to false for the environment in question.
    3. Use `terraform import` from the new metadata module to import the
       existing running infrastructure into the metadata module's state
       file.
    4. Run terraform from the new metadata pipeline to make changes in
i      the usual way.